Update Heroku installation instructions
The current version of CrewLink does not accept URLs if you remove the `https://` from the Heroku URL. The trailing `/` works with or without. Removed reference to manual installation instructions as that talks about server's IP and port, which may be confusing to users as a URL is provided in the previous step.